requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
idna==2.9
IMDbPY==6.8
lxml==4.5.0
pytz==2019.3
requests==2.23.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
urllib3==1.25.9
